backend/pyproject.toml
[tool.poetry]
name = "signup_api"
version = "0.1.4"
description = "An asynchronous Fast API service for signups and Telegram integration"
authors = ["mbrav <mbrav@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "*"
fastapi-pagination = {extras = ["sqlalchemy"], version = "*"}
pydantic = {extras = ["email", "dotenv"], version = "*"}
uvicorn = {extras = ["standard"], version = "*"}
SQLAlchemy = "*"
asyncpg = "*"
python-jose = {extras = ["cryptography"], version = "*"}
passlib = {extras = ["argon2"], version = "*"}
python-multipart = "*"
httpx = {extras = ["http2"], version = "*"}
aiogram = "*"


[tool.poetry.dev-dependencies]
pytest = "*"
pytest-env = "*"
pytest-asyncio = "*"
flake8 = "*"
autopep8 = "*"

[tool.pytest.ini_options]
minversion = "6.2"
testpaths = [
    "tests/",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
]
addopts = "-vv -p no:cacheprovider"
asyncio_mode = "auto"
env = [
    "TESTING=True",
    "LOGGING=False"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
